+++ examples/sample_app/config/routes.rb
@@ -9,6 +9,8 @@ SampleApp::Application.routes.draw do
       resources :authors, only: %i[index show]
       resources :tags, only: %i[index]
       get "authors/by-slug/:slug", to: "authors#by_slug", as: :author_by_slug
+      get "me", to: "me#show"
+      post "login", to: "sessions#create"
       namespace :admin do
         get "stats", to: "stats#show"
       end